+++ src/Node.js
@@ -1156,7 +1156,9 @@
         moveTo: function(newContainer) {
             // do nothing if new container is already parent
             if (this.getParent() !== newContainer) {
-                this.remove();
+                // this.remove my be overrided by drag and drop
+                // buy we need original
+                (this.__originalRemove || this.remove).call(this);
                 newContainer.add(this);
             }
             return this;